Fix order in which we save entities
Looks like we have a problem in our dev stage. On localhost everything
behaves as expected and Hibernate returns proper wrappers for the same
entities. In dev somehow we get different object instances for the same
entity and update the consent status only in of of them. Depending on
the save order we lose updates. Now we fetch the root entity and get the
childs directly from it instead of going through the repo. So same
object instances are guaranteed. Also, because in XS2A everything seems
to use CascadeType.ALL, it's enough to just save the root.